+++ overlays/i3-ws.nix
@@ -1,16 +1,14 @@
 self: super: {
   i3-ws = assert !super ? i3-ws;
-    super.stdenv.mkDerivation (let
+    super.stdenv.mkDerivation (finalAttrs: {
       pname = "i3-ws";
       version = "git-2017-07-30";
-    in {
-      inherit pname version;
 
       src = super.fetchFromGitHub {
         owner = "Enzime";
-        repo = pname;
+        repo = finalAttrs.pname;
         rev = "bca34b6b10509088ceac03fb9a1ef27808165ccb";
-        hash = "1i014hsi80wsaxmdpl8ccwhpvnclkl14w4r0mbci3yqap8a32wgn";
+        hash = "sha256-9nExFLoK+xHZqiATTgKdlNl9IWcM0dtqV5oDFDUkAcQ=";
         fetchSubmodules = true;
       };
 
@@ -20,6 +18,6 @@ self: super: {
       nativeBuildInputs =
         builtins.attrValues { inherit (super) cmake pkg-config; };
 
-      meta.mainProgram = pname;
+      meta.mainProgram = finalAttrs.pname;
     });
 }